Removed collaborator request.
