+++ repository/repository-core/src/main/java/org/eclipse/vorto/repository/web/api/v1/dto/Collaborator.java
@@ -20,25 +20,30 @@ public class Collaborator {
   private String userId;
   private String authenticationProviderId;  
   private String subject;
+  private boolean isTechnicalUser;
   private Collection<String> roles;
   
   public static Collaborator fromUser(TenantUser tenantUser) {
     return new Collaborator(tenantUser.getUser().getUsername(), 
         tenantUser.getUser().getAuthenticationProviderId(),
         tenantUser.getUser().getSubject(),
+        tenantUser.getUser().isTechnicalUser(),
         tenantUser.getRoles().stream().map(role -> role.getRole().name())
           .collect(Collectors.toList()));
   }
   
-  public Collaborator() {
-    
-  }
+  public Collaborator() {}
   
   public Collaborator(String userId, String providerId, String subject, Collection<String> roles) {
+    this(userId, providerId, subject, false, roles);
+  }
+  
+  public Collaborator(String userId, String providerId, String subject, boolean isTechnicalUser, Collection<String> roles) {
     this.userId = userId;
     this.authenticationProviderId = providerId;
     this.roles = roles;
     this.subject = subject;
+    this.isTechnicalUser = isTechnicalUser;
   }
 
   public String getUserId() {
@@ -71,5 +76,22 @@ public class Collaborator {
 
   public void setSubject(String subject) {
     this.subject = subject;
+  }
+  
+  public boolean isTechnicalUser() {
+    return isTechnicalUser;
+  }
+
+  public void setTechnicalUser(boolean isTechnicalUser) {
+    this.isTechnicalUser = isTechnicalUser;
+  }
+  
+  // Jackson mapping oddity
+  public boolean getIsTechnicalUser() {
+    return isTechnicalUser;
+  }
+  
+  public void setIsTechnicalUser(boolean isTechnicalUser) {
+    this.isTechnicalUser = isTechnicalUser;
   }
 }
